data_synthesis_pipeline/swe-scale/requirements.txt
docker
tqdm
unidiff
fastcore
ghapi
beautifulsoup4
python-dotenv
datasets
pyarrow
tenacity
modal
libcst
litellm
astor
pysimdjson
orjson
madoka
beautifulsoup4
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-c-sharp==0.23.1
tree-sitter-cpp==0.23.4
tree-sitter-go==0.25.0
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-php==0.24.1
tree-sitter-python==0.25.0
tree-sitter-ruby==0.23.1
tree-sitter-rust==0.24.0
tree-sitter-typescript==0.23.2
pebble==5.1.3
timeout-decorator


# git clone https://github.com/baidubce/kodo
# cd kodo
# pip install -e .
